pyproject.toml
[tool.poetry]
name = "icalpregnancy"
version = "0.1.0"
description = ""
authors = ["Nadav Angel <nangel@parallelwireless.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
icalendar = "^5.0.11"
flask = "^3.0.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
